git-filter-repo.txt: add Sensitive Data Removal subsection
The DISCUSSION section had lots of details about how to coordinate with
other users when doing a history rewrite.  However, the details were all
in the form of discouraging users from pushing back over a central repo
and instead pushing to a new repo.  For sensitive data removals, just
creating a new repo isn't an option since that leaves the old repo with
the sensitive data exposed.  Add a new section specifically talking
about this involved usecase, and modify the DISCUSSION section to
accomodate the new subsection.

Signed-off-by: Elijah Newren <newren@gmail.com>