filter-repo: avoid parsing the repository refs repeatedly
In commit 704e25878ce1 (filter-repo: limit searches for ref values to
actual refs, 2024-10-21), we added code to be more careful to only
look for ref values corresponding to actual repository refs, but in
doing so, added a bug where we'd repeatedly invoke `git show-ref`
and parse the output, once for every updated ref.  This could be
quite slow, and was entirely unnecessary.  Part of the code was in
place to avoid this, but the key missing setting was overlooked.

Signed-off-by: Elijah Newren <newren@gmail.com>